%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.data.frame.SpotPosition}
\alias{as.data.frame.SpotPosition}
\alias{SpotPosition.as.data.frame}
\alias{as.data.frame.SpotPosition}
\alias{as.data.frame,SpotPosition-method}


\title{Gets a data frame represenation of the spot positions}

\description{
 Gets a data frame represenation of the spot positions.
}

\usage{\method{as.data.frame}{SpotPosition}(x, ...)}

\value{
 Returns a \code{\link[base]{data.frame}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
