%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SMA.R
% 
% on Tue Jan 15 18:37:08 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.layout.SMA}
\alias{as.layout.SMA}
\alias{SMA.as.layout}
\alias{as.layout.SMA}
\alias{as.layout,SMA-method}


\title{Converts a aroma object into an object of the layout structure}

\synopsis{as.layout.SMA(this, obj, ...)}
\usage{SMA$as.layout(obj, ...)}

\description{
  Converts a aroma object into an object of the layout structure, which is used by most of the functions in the sma package.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SMA}}.
}
\keyword{internal}
\keyword{methods}
