%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SMA.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.RG.SMA}
\alias{as.RG.SMA}
\alias{SMA.as.RG}
\alias{as.RG.SMA}
\alias{as.RG,SMA-method}


\title{Converts a aroma object into an object of the RG structure}

\synopsis{as.RG.SMA(this, obj, ...)}
\usage{SMA$as.RG(obj, ...)}

\description{
  Converts a aroma object into an object of the RG structure, which is used by most of the function in the sma package.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SMA}}.
}
\keyword{internal}
\keyword{methods}
