%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% on Tue Jan 15 18:37:36 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAspectRatio.SpotPosition}
\alias{getAspectRatio.SpotPosition}
\alias{SpotPosition.getAspectRatio}
\alias{getAspectRatio.SpotPosition}
\alias{getAspectRatio,SpotPosition-method}


\title{Gets the aspect ratio of the spots}

\usage{\method{getAspectRatio}{SpotPosition}(this, ...)}

\description{
  Gets the aspect ratio of the spots, i.e. the ratio between the maximum height and the maximum
  width. If all spots on an array are included this is \emph{approximately}
  the aspect ratio of the array itself. This is useful to make graphs that
  display true physical positions.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
