%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getBottomEdge.SpotPosition}
\alias{getBottomEdge.SpotPosition}
\alias{SpotPosition.getBottomEdge}
\alias{getBottomEdge.SpotPosition}
\alias{getBottomEdge,SpotPosition-method}


\title{Gets the y-coordinate of the spot at the very bottom}

\usage{\method{getBottomEdge}{SpotPosition}(this, ...)}

\description{
  Gets the y-coordinate of the spot at the very bottom.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
