%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% on Tue Jan 15 18:37:36 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDistancesTo.SpotPosition}
\alias{getDistancesTo.SpotPosition}
\alias{SpotPosition.getDistancesTo}
\alias{getDistancesTo.SpotPosition}
\alias{getDistancesTo,SpotPosition-method}


\title{Gets the Euclidean distance from one spot to a set of other spots}

\usage{\method{getDistancesTo}{SpotPosition}(this, index, ...)}

\description{
  Gets the Euclidean distance from one spot to a set of other spots.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
