%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFieldNames.MicroarrayData}
\alias{getFieldNames.MicroarrayData}
\alias{MicroarrayData.getFieldNames}
\alias{getFieldNames.MicroarrayData}
\alias{getFieldNames,MicroarrayData-method}


\title{Gets the names of all fields containing spot specific data}

\usage{\method{getFieldNames}{MicroarrayData}(this, ...)}

\description{
 Gets the names of all fields containing spot specific data.
 This is a read-only method. Thus, there is no corresponding set-method.
 It is used mainly for internal purposes.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
