%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  AcceptFilter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getIndex.AcceptFilter}
\alias{getIndex.AcceptFilter}
\alias{AcceptFilter.getIndex}
\alias{getIndex.AcceptFilter}
\alias{getIndex,AcceptFilter-method}


\title{Gets indices accepted by this filter}

\usage{\method{getIndex}{AcceptFilter}(this, ...)}

\description{
 Gets indices accepted by this filter.
}

\value{
  Returns a \code{\link[base]{vector}} of \code{\link[base]{integer}}s.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{AcceptFilter}}..
}
\keyword{internal}
\keyword{methods}
