%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FieldFilter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getIndex.FieldFilter}
\alias{getIndex.FieldFilter}
\alias{FieldFilter.getIndex}
\alias{getIndex.FieldFilter}
\alias{getIndex,FieldFilter-method}


\title{Gets indices accepted by this filter}

\usage{\method{getIndex}{FieldFilter}(this, ...)}

\description{
 .
}

\value{
  Returns a \code{\link[base]{vector}} of \code{\link[base]{integer}}s.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{FieldFilter}}..
}
\keyword{internal}
\keyword{methods}
