%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Filter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getIndex.Filter}
\alias{getIndex.Filter}
\alias{Filter.getIndex}
\alias{getIndex.Filter}
\alias{getIndex,Filter-method}


\title{Gets the indices accepted by the filter}

\usage{\method{getIndex}{Filter}(...)}

\description{
  Gets the indices that this filter passes through. The filter asks
  all its input connections for indices and passes through only those
  indices that are accepted by this filter.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{Filter}}..
}
\keyword{internal}
\keyword{methods}
