%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Filter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInput.Filter}
\alias{getInput.Filter}
\alias{Filter.getInput}
\alias{getInput.Filter}
\alias{getInput,Filter-method}


\title{Gets all the input objects connected to the filter}

\usage{\method{getInput}{Filter}(...)}

\description{
  Gets all the input objects connected to the filter.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  To change one or more input objects see \code{\link[aroma:changeInput.Filter]{*changeInput}()}.
  For more information see \code{\link{Filter}}..
}
\keyword{internal}
\keyword{methods}
