%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ParallelFilter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInput.ParallelFilter}
\alias{getInput.ParallelFilter}
\alias{ParallelFilter.getInput}
\alias{getInput.ParallelFilter}
\alias{getInput,ParallelFilter-method}


\title{Gets the input filters}

\usage{\method{getInput}{ParallelFilter}(this, ...)}

\description{
 Gets the input filters.
}

\value{
  Returns a \code{\link[base]{list}} of \code{\link{Filter}} objects.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{ParallelFilter}}..
}
\keyword{internal}
\keyword{methods}
