%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SerialFilter.R
% 
% on Tue Jan 15 18:37:04 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getInput.SerialFilter}
\alias{getInput.SerialFilter}
\alias{SerialFilter.getInput}
\alias{getInput.SerialFilter}
\alias{getInput,SerialFilter-method}


\title{Gets all the input objects connected to the filter}

\usage{\method{getInput}{SerialFilter}(this, ...)}

\description{
  Gets all the input objects connected to the filter.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  To change one or more input objects see \code{\link{Filter.changeInput}}.
  For more information see \code{\link{SerialFilter}}.
}
\keyword{internal}
\keyword{methods}
