%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% on Tue Jan 15 18:37:36 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getMaxHeight.SpotPosition}
\alias{getMaxHeight.SpotPosition}
\alias{SpotPosition.getMaxHeight}
\alias{getMaxHeight.SpotPosition}
\alias{getMaxHeight,SpotPosition-method}


\title{Gets the maximum y-distance between two spots}

\usage{\method{getMaxHeight}{SpotPosition}(this, ...)}

\description{
  Gets the maximum y-distance between two spots.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
