%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getMaxWidth.SpotPosition}
\alias{getMaxWidth.SpotPosition}
\alias{SpotPosition.getMaxWidth}
\alias{getMaxWidth.SpotPosition}
\alias{getMaxWidth,SpotPosition-method}


\title{Gets the maximum x-distance between two spots}

\usage{\method{getMaxWidth}{SpotPosition}(this, ...)}

\description{
  Gets the maximum x-distance between two spots.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
