%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GeneGroups.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getNames.GeneGroups}
\alias{getNames.GeneGroups}
\alias{GeneGroups.getNames}
\alias{getNames.GeneGroups}
\alias{getNames,GeneGroups-method}


\title{Gets the names of the unique gene names}

\usage{\method{getNames}{GeneGroups}(this, ...)}

\description{
 Gets the names of the unique gene names.
}

\value{
  Returns a \code{\link[base]{vector}} of length \code{nbrOfGroups(this)}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GeneGroups}}.
}
\keyword{internal}
\keyword{methods}
