%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% on Tue Jan 15 18:37:36 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getRightEdge.SpotPosition}
\alias{getRightEdge.SpotPosition}
\alias{SpotPosition.getRightEdge}
\alias{getRightEdge.SpotPosition}
\alias{getRightEdge,SpotPosition-method}


\title{Gets the x-coordinate of the right most spot}

\usage{\method{getRightEdge}{SpotPosition}(this, ...)}

\description{
  Gets the x-coordinate of the right most spot.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
