%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GeneGroups.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSizes.GeneGroups}
\alias{getSizes.GeneGroups}
\alias{GeneGroups.getSizes}
\alias{getSizes.GeneGroups}
\alias{getSizes,GeneGroups-method}


\title{Gets the number of replicates for each gene}

\usage{\method{getSizes}{GeneGroups}(this, ...)}

\description{
 Gets the number of replicates for each gene.
}

\value{
  Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} of length \code{nbrOfGroups(this)}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GeneGroups}}.
}
\keyword{internal}
\keyword{methods}
