%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getTopEdge.SpotPosition}
\alias{getTopEdge.SpotPosition}
\alias{SpotPosition.getTopEdge}
\alias{getTopEdge.SpotPosition}
\alias{getTopEdge,SpotPosition-method}


\title{Gets the y-coordinate of the spot at the very top}

\usage{\method{getTopEdge}{SpotPosition}(this, ...)}

\description{
  Gets the y-coordinate of the spot at the very top.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
