%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.R
% 
% on Tue Jan 15 18:34:03 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{listFlags.MicroarrayData}
\alias{listFlags.MicroarrayData}
\alias{MicroarrayData.listFlags}
\alias{listFlags.MicroarrayData}
\alias{listFlags,MicroarrayData-method}


\title{Lists the names of all known flags}

\description{
  Lists the names of all known flags that matches a regular expression.
}

\usage{\method{listFlags}{MicroarrayData}(this, regexpr=NULL, ...)}

\arguments{
 \item{regexpr}{The regular expression of which flags should be cleared.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:getFlag.MicroarrayData]{*getFlag}()},
  \code{\link[aroma:setFlag.MicroarrayData]{*setFlag}()},
  \code{\link[aroma:clearFlag.MicroarrayData]{*clearFlag}()},
  \code{\link[aroma:getInclude.MicroarrayData]{*getInclude}()}.
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
