%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GeneGroups.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfGroups.GeneGroups}
\alias{nbrOfGroups.GeneGroups}
\alias{GeneGroups.nbrOfGroups}
\alias{nbrOfGroups.GeneGroups}
\alias{nbrOfGroups,GeneGroups-method}


\title{Gets the number of unique genes}

\usage{\method{nbrOfGroups}{GeneGroups}(this, ...)}

\description{
 Gets the number of unique genes.
}

\value{
  Returns an \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GeneGroups}}.
}
\keyword{internal}
\keyword{methods}
