%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfSlides.MicroarrayData}
\alias{nbrOfSlides.MicroarrayData}
\alias{MicroarrayData.nbrOfSlides}
\alias{nbrOfSlides.MicroarrayData}
\alias{nbrOfSlides,MicroarrayData-method}


\title{Gets the number of slides}

\usage{\method{nbrOfSlides}{MicroarrayData}(this, ...)}

\description{
  Gets the number of slides.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:nbrOfFields.MicroarrayData]{*nbrOfFields}()},
  \code{\link[aroma:nbrOfSpots.MicroarrayData]{*nbrOfSpots}()},
  \code{\link[aroma:size.MicroarrayData]{*size}()}.
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
