%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfSlides.SpotPosition}
\alias{nbrOfSlides.SpotPosition}
\alias{SpotPosition.nbrOfSlides}
\alias{nbrOfSlides.SpotPosition}
\alias{nbrOfSlides,SpotPosition-method}


\title{Gets the number of slides}

\usage{\method{nbrOfSlides}{SpotPosition}(this, ...)}

\description{
  Gets the number of slides.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:nbrOfSpots.SpotPosition]{*nbrOfSpots}()},
  \code{\link[aroma:size.SpotPosition]{*size}()}.
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
