%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfSpots.MicroarrayData}
\alias{nbrOfSpots.MicroarrayData}
\alias{MicroarrayData.nbrOfSpots}
\alias{nbrOfSpots.MicroarrayData}
\alias{nbrOfSpots,MicroarrayData-method}


\title{Gets the number of spots in each of the slides}

\usage{\method{nbrOfSpots}{MicroarrayData}(this, ...)}

\description{
  Gets the number of spots in each of the slides.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:nbrOfFields.MicroarrayData]{*nbrOfFields}()},
  \code{\link[aroma:nbrOfSlides.MicroarrayData]{*nbrOfSlides}()},
  \code{\link[aroma:size.MicroarrayData]{*size}()}.
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
