%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfSpots.SpotPosition}
\alias{nbrOfSpots.SpotPosition}
\alias{SpotPosition.nbrOfSpots}
\alias{nbrOfSpots.SpotPosition}
\alias{nbrOfSpots,SpotPosition-method}


\title{Gets the number of spots in each of the slides}

\usage{\method{nbrOfSpots}{SpotPosition}(this, ...)}

\description{
  Gets the number of spots in each of the slides.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:nbrOfSlides.SpotPosition]{*nbrOfSlides}()},
  \code{\link[aroma:size.SpotPosition]{*size}()}.
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
