%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{plot.SpotPosition}
\alias{plot.SpotPosition}
\alias{SpotPosition.plot}
\alias{plot.SpotPosition}
\alias{plot,SpotPosition-method}


\title{Plots the spot positions in an xy scatter plot}

\usage{\method{plot}{SpotPosition}(x, xlab="x", ylab="y", pch=20, ...)}

\description{
  Plots the spot positions in an xy scatter plot.
}

\arguments{
  \item{xlab, ylab}{\code{\link[base]{character}} string or \code{\link[base]{expression}} specifying the
      x and the y labels.}
  \item{pch}{The point character to be plotted at each position.}
  \item{...}{All other arguments accepted by the \code{plot()} function
      found in the base package.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:points.SpotPosition]{*points}()}
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
