%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% on Tue Jan 15 18:37:36 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{points.SpotPosition}
\alias{points.SpotPosition}
\alias{SpotPosition.points}
\alias{points.SpotPosition}
\alias{points,SpotPosition-method}


\title{Adds spot positions to an existing plot}

\usage{\method{points}{SpotPosition}(x, index=NULL, pch=20, ...)}

\description{
  Adds spot positions to an existing plot.
}

\arguments{
  \item{pch}{The point character to be plotted at each position.}
  \item{...}{All other arguments accepted by the \code{points()} function
      found in the base package.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:plot.SpotPosition]{*plot}()}
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
