%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.IO.R
% 
% on Tue Jan 15 18:33:27 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{read.MicroarrayData}
\alias{read.MicroarrayData}
\alias{MicroarrayData.read}
\alias{read.MicroarrayData}
\alias{read,MicroarrayData-method}


\title{Reads microarray data generated by another software package}

\synopsis{read.MicroarrayData(static, ..., verbose=FALSE)}
\usage{MicroarrayData$read(..., verbose=FALSE)}

\arguments{
  \item{...}{Any arguments accepted by the read() methods of the
   subclasses. Such arguments are commonly: filename, pattern and path.}
  \item{verbose}{If \code{\link[base:logical]{TRUE}}, information will printed out during
                 the reading of the file.}
}

\value{
  Returns a \code{\link{MicroarrayData}} object.
}

\description{
  Static method that reads microarray data generated by another software
  package into a MicroarrayData object. If the a filename pattern is
  used, the files are guaranteed to be read in lexical (alphabetic)
  order.
}

\examples{
  \dontrun{
  # Due to a bug in R CMD check (R v1.7.1) the MicroarrayData$read() call
  # below will call getKnownSubclasses(), which will generate
  #   "Error in exists(objectName, mode = "function") :
  #	   [2003-07-07 23:32:41] Exception: F used instead of FALSE"
  # Note that the example still work, just not in R CMD check

  sa <- MicroarrayData$read(pattern="group.*.dat", path=system.file("data-ex", package="aroma"))
  }
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{internal}
\keyword{methods}
