%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ScanAlyze20Data.R
% 
% on Tue Jan 15 18:36:41 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{read.ScanAlyze20Data}
\alias{read.ScanAlyze20Data}
\alias{ScanAlyze20Data.read}
\alias{read.ScanAlyze20Data}
\alias{read,ScanAlyze20Data-method}


\title{Reads one or several ScanAlyze v2.0 files}

\description{
 Reads one or several ScanAlyze v2.0 files.
}

\synopsis{read.ScanAlyze20Data(static, ...)}
\usage{ScanAlyze20Data$read(...)}

\arguments{
  \item{...}{Same arguments as read() in ScanAlyzeData class.}
}

\value{Returns a \code{\link{ScanAlyze20Data}} object.}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\references{
  \url{http://array.mbb.yale.edu/analysis/manual/inputfiles/}
}

\seealso{
  For pattern formats see \code{\link[base]{list.files}}().
  \code{\link{ScanAlyzeData}}.
  For more information see \code{\link{ScanAlyze20Data}}.
}
\keyword{internal}
\keyword{methods}
