%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotfinderData.R
% 
% on Tue Jan 15 18:37:29 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readOneFile.SpotfinderData}
\alias{readOneFile.SpotfinderData}
\alias{SpotfinderData.readOneFile}
\alias{readOneFile.SpotfinderData}
\alias{readOneFile,SpotfinderData-method}


\title{Reads TAV files generated by Spotfinder}

\description{
 Reads TAV files generated by Spotfinder.
}

\synopsis{readOneFile.SpotfinderData(this, filename, path=NULL, sep="auto", skip="auto", verbose=TRUE, ...)}
\usage{SpotfinderData$readOneFile(filename, path=NULL, sep="auto", skip="auto", verbose=TRUE, ...)}

\arguments{
  \item{filenames}{The filename(s) of the TAV/Spotfinder file(s) to be read.}
  \item{path}{The path(s) to the file(s).}
  \item{sep}{Character that separates the columns of data.
    If \code{"auto"} a best guess is made from information in the header.}
  \item{skip}{Number of lines to skip \emph{before} reading the header.
    If \code{"auto"}, the number of lines to skip before the header will
    be searched for automatically.}
  \item{verbose}{If \code{\link[base:logical]{TRUE}}, information will printed out during
                 the file reading/parsing.}
}

\value{
 Returns a \code{\link{SpotfinderData}} object containing one or several slides.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\examples{\dontrun{# For an example see help(SpotfinderData).}}

\seealso{
  For more information see \code{\link{SpotfinderData}}.
}
\keyword{internal}
\keyword{methods}
