%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{removeSlides.MicroarrayData}
\alias{removeSlides.MicroarrayData}
\alias{MicroarrayData.removeSlides}
\alias{removeSlides.MicroarrayData}
\alias{removeSlides,MicroarrayData-method}


\title{Remove specified slides}

\description{
  Remove specified slides.
}

\usage{\method{removeSlides}{MicroarrayData}(this, slides, ...)}

\arguments{
 \item{slides}{\code{\link[base]{vector}} of slides to be removed.
   If \code{\link[base]{NULL}}, all slides are removed.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:append.MicroarrayData]{*append}()} and \code{\link[aroma:keepSlides.MicroarrayData]{*keepSlides}()}.
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
