%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Filter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setParameter.Filter}
\alias{setParameter.Filter}
\alias{Filter.setParameter}
\alias{setParameter.Filter}
\alias{setParameter,Filter-method}


\title{Sets the values for a specific parameter for indices accepted by the filter}

\usage{\method{setParameter}{Filter}(this, paramName, value, ...)}

\description{
  Sets the values for a specific parameter for those indices that the
  filter lets through. For more information see \code{\link[aroma:getParameter.Filter]{*getParameter}()}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  To get a parameter of a filter see \code{\link[aroma:getParameter.Filter]{*getParameter}()}.
  To get the indices accepted by a filter see \code{\link[aroma:getIndex.Filter]{*getIndex}()}.
  For more information see \code{\link{Filter}}..
}
\keyword{internal}
\keyword{methods}
