%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Filter.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setVisible.Filter}
\alias{setVisible.Filter}
\alias{Filter.setVisible}
\alias{setVisible.Filter}
\alias{setVisible,Filter-method}


\title{Sets if the filter should bee visible or not}

\usage{\method{setVisible}{Filter}(obj, visible, ...)}

\description{
  Sets if the filter should bee visible or not.
  By default a filter is visible.
}

\arguments{
  \item{visible}{Logical value specifying if the filter should be
    visible or not.}
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:getVisible.Filter]{*getVisible}()}.
  For more information see \code{\link{Filter}}..
}
\keyword{internal}
\keyword{methods}
