%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{size.SpotPosition}
\alias{size.SpotPosition}
\alias{SpotPosition.size}
\alias{size.SpotPosition}
\alias{size,SpotPosition-method}


\title{Gets the number of spots in all slides together}

\usage{\method{size}{SpotPosition}(this, ...)}

\description{
  Gets the number of spots in all slides together.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:nbrOfSlides.SpotPosition]{*nbrOfSlides}()},
  \code{\link[aroma:nbrOfSpots.SpotPosition]{*nbrOfSpots}()}.
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
