%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SMA.R
% 
% on Tue Jan 15 18:37:08 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{SMA}
\docType{class}
\alias{SMA}


\title{The SMA class}

\description{
 Package:  aroma \cr
\bold{Class SMA}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{SMA}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{SMA}\cr
extends \link[R.oo]{Object}\cr



  The SMA class is a static class that provided methods to convert
  data object in aroma into sma structures. This could be
  useful if you would like to use methods in sma that are not (yet)
  implemented in the aroma package.
}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[aroma:as.layout.SMA]{as.layout}} \tab Converts a aroma object into an object of the layout structure.\cr
 \tab \code{\link[aroma:as.RG.SMA]{as.RG}} \tab Converts a aroma object into an object of the RG structure.\cr
 \tab \code{loadData} \tab  -\cr
 \tab \code{unloadData} \tab  -\cr
}


 \bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstanciationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
