%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{str.MicroarrayData}
\alias{str.MicroarrayData}
\alias{MicroarrayData.str}
\alias{str.MicroarrayData}
\alias{str,MicroarrayData-method}


\title{Compactly Display the Structure of a MicroarrayData object}

\description{
  Compactly Display the Structure of a MicroarrayData object.
}

\usage{\method{str}{MicroarrayData}(object, ...)}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[base]{str}}.
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
