%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{summary.MicroarrayData}
\alias{summary.MicroarrayData}
\alias{MicroarrayData.summary}
\alias{summary.MicroarrayData}
\alias{summary,MicroarrayData-method}


\title{Gets summary statistics for the fields}

\description{
  Gets summary statistics for the fields.
  This is identical to calling \code{\link[base]{summary}}() on the result from
  \code{\link[aroma:as.data.frame.MicroarrayData]{*as.data.frame}()}.
}

\usage{\method{summary}{MicroarrayData}(object, ...)}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:as.data.frame.MicroarrayData]{*as.data.frame}()}.
  \code{\link[base]{summary}}().
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
