%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.ARGS.R
% 
% on Tue Jan 15 18:33:27 2008.
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{validateArgumentSlide.MicroarrayData}
\alias{validateArgumentSlide.MicroarrayData}
\alias{MicroarrayData.validateArgumentSlide}
\alias{validateArgumentSlide.MicroarrayData}
\alias{validateArgumentSlide,MicroarrayData-method}


\title{Validates the argument 'slide'}

\description{
 Validates the argument 'slide'.
}

\usage{\method{validateArgumentSlide}{MicroarrayData}(this, slide=NULL, ...)}

\arguments{
  \item{slide}{The argument value to be validated.}
}

\value{
 Returns an interpreted and valid value.
 If the argument was invalid an \code{\link[R.oo]{Exception}} is thrown.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
