%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  SpotPosition.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.SpotPosition}
\alias{as.character.SpotPosition}
\alias{SpotPosition.as.character}
\alias{as.character.SpotPosition}
\alias{as.character,SpotPosition-method}


\title{Gets a character represenation of the spot positions}

\description{
 Gets a character represenation of the spot positions.
}

\usage{\method{as.character}{SpotPosition}(this, ...)}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{SpotPosition}}.
}
\keyword{internal}
\keyword{methods}
