%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MicroarrayData.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{size.MicroarrayData}
\alias{size.MicroarrayData}
\alias{MicroarrayData.size}
\alias{size.MicroarrayData}
\alias{size,MicroarrayData-method}


\title{Gets the number of spots in all slides together}

\usage{\method{size}{MicroarrayData}(this, ...)}

\description{
  Gets the number of spots in all slides together.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[aroma:nbrOfFields.MicroarrayData]{*nbrOfFields}()},
  \code{\link[aroma:nbrOfSlides.MicroarrayData]{*nbrOfSlides}()},
  \code{\link[aroma:nbrOfSpots.MicroarrayData]{*nbrOfSpots}()}.
  For more information see \code{\link{MicroarrayData}}.
}
\keyword{internal}
\keyword{methods}
